fix bug

+++ app/asset/routes.py
@@ -638,25 +638,30 @@ def UploadReport(id):
     form = UploadReportForm(obj=workorder)
     role = get_userrole(current_user.id)
     if form.validate_on_submit():
-        workorder.status = 1
-        transaction = Production(wo=form.wo.data, pn=form.pn.data, csn=form.csn.data, msn=form.msn.data, cpu=form.cpu.data, 
-        mem1=form.mem1.data, mem2=form.mem2.data, mem3=form.mem3.data, mem4=form.mem4.data, gpu1=form.gpu1.data, gpu2=form.gpu2.data, sata1=form.sata1.data, sata2=form.sata2.data,
-        sata3=form.sata3.data, sata4=form.sata4.data, m21=form.m21.data, m22=form.m22.data, wifi=form.wifi.data, fg5g=form.fg5g.data,
-        can=form.can.data,other=form.other.data,note=form.note.data,report=form.report.data)
-        existproduct = Production.query.filter_by(wo=form.wo.data,csn=form.csn.data.strip())
-        if existproduct.count() :
-            db.session.delete(existproduct[0])
-        db.session.add(transaction)
-        db.session.commit()
-        fstr = "Upload successful! "
-        if workorder.cpuinstall == False and ("Nuvo" in workorder.pn or "SEMIL" in workorder.pn):
-            fstr += "PLEASE uninstall CPU. "
-        if workorder.memoryinstall == False and ("Nuvo" in workorder.pn or "POC" in workorder.pn or "SEMIL" in workorder.pn) :    
-            fstr += "PLEASE uninstall Memory."
-        if "WARNING" in form.note.data :
-            fstr += form.note.data     
-        flash(fstr)
-        return redirect(url_for('main.display_workorders'))
+        if(workorder.status != 0):
+            fstr = "Please take workorder before upload report!"
+            flash(fstr)
+            return render_template('uploadreport.html', form=form, id=id,userrole = role)
+        else :     
+            workorder.status = 1
+            transaction = Production(wo=form.wo.data, pn=form.pn.data, csn=form.csn.data, msn=form.msn.data, cpu=form.cpu.data, 
+            mem1=form.mem1.data, mem2=form.mem2.data, mem3=form.mem3.data, mem4=form.mem4.data, gpu1=form.gpu1.data, gpu2=form.gpu2.data, sata1=form.sata1.data, sata2=form.sata2.data,
+            sata3=form.sata3.data, sata4=form.sata4.data, m21=form.m21.data, m22=form.m22.data, wifi=form.wifi.data, fg5g=form.fg5g.data,
+            can=form.can.data,other=form.other.data,note=form.note.data,report=form.report.data)
+            existproduct = Production.query.filter_by(wo=form.wo.data,csn=form.csn.data.strip())
+            if existproduct.count() :
+                db.session.delete(existproduct[0])
+            db.session.add(transaction)
+            db.session.commit()
+            fstr = "Upload successful! "
+            if workorder.cpuinstall == False and ("Nuvo" in workorder.pn or "SEMIL" in workorder.pn):
+                fstr += "PLEASE uninstall CPU. "
+            if workorder.memoryinstall == False and ("Nuvo" in workorder.pn or "POC" in workorder.pn or "SEMIL" in workorder.pn) :    
+                fstr += "PLEASE uninstall Memory."
+            if "WARNING" in form.note.data :
+                fstr += form.note.data     
+            flash(fstr)
+            return redirect(url_for('main.display_workorders'))
     return render_template('uploadreport.html', form=form, id=id,userrole = role)
 
 @main.route('/ReviewReport/<id>', methods=['GET', 'POST'])
